+++ cli/src/lib.rs
@@ -49,7 +49,7 @@ pub fn get_entries_for_profile(
 pub fn write_target(
     entries: Vec<String>,
     target_file_path: &String,
-    region_output_name: &String
+    region_name_override: &String
 ) -> Result<(), Box<dyn std::error::Error>> {
     let file = File::create(target_file_path)?;
     let mut writer = BufWriter::new(file);
@@ -63,7 +63,7 @@ pub fn write_target(
                     writeln!(&mut w, "{} = \"{}\"", k.trim(), v.trim())?;
                 }
                 "region" => {
-                    writeln!(&mut w, "{} = \"{}\"", region_output_name.trim(), v.trim())?;
+                    writeln!(&mut w, "{} = \"{}\"", region_name_override.trim(), v.trim())?;
                 }
                 _ => (),
             }
@@ -78,10 +78,10 @@ pub fn parse_and_write(
     source_file_path: &String,
     target_file_path: &String,
     profile: &String,
-    region_output_name: &String
+    region_name_override: &String
 ) -> Result<(), Box<dyn Error>> {
     let source_map = parse_source(source_file_path).unwrap();
     let entries = get_entries_for_profile(source_map, profile)?;
-    write_target(entries, target_file_path, region_output_name)?;
+    write_target(entries, target_file_path, region_name_override)?;
     Ok(())
 }